requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
faiss-cpu
openai
-U langchain-community
tiktoken